Fix for issue #2225 (#2233)
* Fix for issue #2225

Added a check for file users/default.png (default avatar image) when deleting image files.

* Fix to pass StyleCI check

* Fix to pass StyleCI test

* Updated package.json to latest versions

* Revert "Updated package.json to latest versions"

This reverts commit bc2b179a1c8573eec8766e1ed1590ed98b33d9d5.

* Changed hardcoded filename with config variable

+++ src/Http/Controllers/VoyagerBreadController.php
@@ -423,7 +423,9 @@ class VoyagerBreadController extends Controller
     public function deleteBreadImages($data, $rows)
     {
         foreach ($rows as $row) {
-            $this->deleteFileIfExists($data->{$row->field});
+            if ($data->{$row->field} != config('voyager.user.default_avatar')) {
+                $this->deleteFileIfExists($data->{$row->field});
+            }
 
             $options = json_decode($row->details);
